Qwen2_ColPali/requirements.txt
byaldi 
pdf2image 
qwen-vl-utils 
transformers
torchvision
opencv-python
PyMuPDF